added following and followers to sidebar as well as refactored into single table component for both
