--- client/src/app/reducers/selectedContentPage.js
@@ -1,19 +1,0 @@
-// import contentType from "../../data/ContentTypeEnum";
-
-// const initialState = contentType.HOME;
-
-// export const selectedContentPage = (state = initialState, action) => {
-//     switch (action.type) {
-//         case contentType.FEED:
-//             return contentType.FEED;
-//         case contentType.PROFILE:
-//             return contentType.PROFILE;
-//         case contentType.HOME:
-//             return contentType.HOME;
-//         default:
-//             return state;
-//     }
-
-// }
-
-// export default selectedContentPage;
