--- client/src/components/Followers.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-class Followers extends React.Component{
-    render() {
-     return <h1>Follower Component Placeholder</h1>
- }
-
-}
-
-export default Followers;
