--- client/src/components/Following.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class Following extends React.Component {
-  render() {
-    return <h1>Following Component Placeholder</h1>;
-  }
-}
-
-export default Following;
