+++ client/src/components/profile/ProfileTable.js
@@ -1,12 +1,13 @@
 import React from "react";
 import { connect } from "react-redux";
 import Spotify from "spotify-web-api-js";
+import FollowTable from "../FollowTable";
 
 import MaterialTable from "material-table";
 import { Paper, Tab, Tabs } from "@material-ui/core";
-import { withStyles, createMuiTheme } from "@material-ui/core/styles";
+import { withStyles } from "@material-ui/core/styles";
 
-const styles = (theme) => ({
+const styles = () => ({
   root: {
     flexGrow: 1,
     padding: 10,
@@ -28,41 +29,13 @@ class ProfileTable extends React.Component {
     spotifyWebApi
       .getUserPlaylists(this.props.user.id)
       .then((result) => {
-        console.log(result);
         const playlists = this.transformPlaylistData(result);
         this.setState({
           playlists: playlists,
         });
       })
       .catch((err) => {
-        console.log("error getting top tracks");
-        console.log(err);
-      });
-    fetch(`http://localhost:9000/user/following/${this.props.user.id}`)
-      .then((res) => res.json())
-      .then((res) => {
-        if (res.error) {
-          throw res.error;
-        }
-        console.log("following res: ", res);
-        const following = this.transformFollowData(res.data[0].following);
-        this.setState({ following: following });
-      })
-      .catch((error) => {
-        console.log("Fetch Following Error: ", error);
-      });
-    fetch(`http://localhost:9000/user/followers/${this.props.user.id}`)
-      .then((res) => res.json())
-      .then((res) => {
-        if (res.error) {
-          throw res.error;
-        }
-        console.log("followers res: ", res);
-        const followers = this.transformFollowData(res.data[0].followers);
-        this.setState({ followers: followers });
-      })
-      .catch((error) => {
-        console.log("Fetch Followers Error: ", error);
+        console.log("Error getting top tracks: ", err);
       });
   };
 
@@ -77,19 +50,6 @@ class ProfileTable extends React.Component {
       return playlist;
     });
     return playlists;
-  };
-
-  transformFollowData = (data) => {
-    const follData = data.map((f) => {
-      const foll = {
-        pic: f.profilePic,
-        username: f.username,
-        numPosts: f.posts.length,
-        numFollowers: f.followers.length,
-      };
-      return foll;
-    });
-    return follData;
   };
 
   handleChange = (event, index) => {
@@ -141,44 +101,9 @@ class ProfileTable extends React.Component {
       );
     } else {
       return (
-        <MaterialTable
-          components={{
-            Container: (props) => (
-              <Paper {...props} elevation={0} style={{ boxShadow: 0 }} />
-            ),
-          }}
-          columns={[
-            {
-              title: "MySpotter",
-              field: "pic",
-              render: (rowData) => (
-                <img
-                  src={rowData.pic}
-                  alt={"ProfilePic"}
-                  style={{ width: 40, height: 40, borderRadius: 16 }}
-                />
-              ),
-              headerStyle: { width: "50px" },
-              cellStyle: { width: "50px" },
-              width: null,
-            },
-            { title: "", field: "username" },
-            { title: "# of Posts", field: "numPosts" },
-            { title: "# of Followers", field: "numFollowers" },
-          ]}
-          data={
-            this.state.tabIndex === 1
-              ? this.state.followers
-              : this.state.following
-          }
-          options={{
-            showTitle: false,
-            search: false,
-            paging: false,
-            toolbar: false,
-            sorting: false,
-            rowStyle: { borderBottom: 0 },
-          }}
+        <FollowTable
+          key={this.state.tabIndex}
+          type={this.state.tabIndex === 1 ? "followers" : "following"}
         />
       );
     }
@@ -210,15 +135,7 @@ class ProfileTable extends React.Component {
 const mapStateToProps = (state) => ({
   spotifyWebApi: state.spotifyWebApi,
   spotifyApiUserMe: state.spotifyApiUserMe,
-  posts: state.posts,
   user: state.user,
 });
 
-const mapDispatchToProps = {
-  // getFollowersAndFollowing,
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(withStyles(styles)(ProfileTable));
+export default connect(mapStateToProps)(withStyles(styles)(ProfileTable));
